requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pyOpenSSL==25.1.0
Requests==2.32.4
tabulate==0.9.0
